Fix sys/win32/* for x86_64

+++ neo/sys/win32/win_main.cpp
@@ -419,7 +419,7 @@ uintptr_t Sys_DLL_Load( const char *dllName ) {
 		GetModuleFileName( libHandle, loadedPath, sizeof( loadedPath ) - 1 );
 		if ( idStr::IcmpPath( dllName, loadedPath ) ) {
 			Sys_Printf( "ERROR: LoadLibrary '%s' wants to load '%s'\n", dllName, loadedPath );
-			Sys_DLL_Unload( (int)libHandle );
+			Sys_DLL_Unload( (uintptr_t)libHandle );
 			return 0;
 		}
 	}
